requirements.txt
discord.py
flask
python_dotenv